Refactor the notary client library to have a separate reader and downloader that may be easier to use for clients that do not need any publishing capabilities.
